requirements.txt
fastapi
pd
pandas
openpyxl
python-multipart
uvicorn
jinja2
python-dotenv
supabase
